Make VS's classes not under game_frame work. Tidy up include path.
