+++ Source/Game/VSclass/Button.cpp
@@ -1,0 +1,23 @@
+#include "stdafx.h"
+#include "../../Library/gameutil.h"
+#include "Button.h"
+Button::Button() 
+{
+}
+Button::~Button()
+{
+}
+void Button::is_hover() 
+{
+}
+void Button::is_press()
+{
+}
+void Button::set_pos(int x, int y)
+{
+	_x = x;
+	_y = y;
+}
+void Button::show_button() 
+{
+}
