+++ Source/Game/VSclass/Map.cpp
@@ -1,0 +1,22 @@
+#include "stdafx.h"
+#include "../../Library/gameutil.h"
+#include "Map.h"
+
+Map::Map()
+{
+}
+Map::~Map() 
+{
+}
+void Map::load_map()
+{
+	this->base_backgrounds.LoadBitmapByString({});
+}
+void Map::show_map()
+{
+	
+}
+void Map::map_padding()
+{
+	
+}
