+++ Source/Game/VSclass/Passive.cpp
@@ -1,0 +1,13 @@
+#include "stdafx.h"
+#include "../../Library/gameutil.h"
+#include "Passive.h"
+Passive::Passive()
+{
+}
+Passive::~Passive()
+{
+}
+void Passive::level_up()
+{
+	this->_level = _level + 1;
+}
