+++ Source/Game/VSclass/Passive.h
@@ -1,0 +1,11 @@
+#pragma once
+class Passive {
+public:
+	Passive();
+	~Passive();
+	void level_up();
+private:
+	int _level;
+	int _type;
+	game_framework::CMovingBitmap _skin;
+};
