+++ Source/Game/VSclass/VS.h
@@ -1,0 +1,8 @@
+#pragma once
+#include "Button.h"
+#include "Enemy.h"
+#include "Map.h"
+#include "Passive.h"
+#include "Weapon.h"
+#include "Pickup.h"
+#include "Player.h"
