+++ Source/Game/VSclass/Weapon.cpp
@@ -1,11 +1,5 @@
 #include "stdafx.h"
-#include "../Core/Resource.h"
-#include <mmsystem.h>
-#include <ddraw.h>
-#include "../Library/audio.h"
-#include "../Library/gameutil.h"
-#include "../Library/gamecore.h"
-#include "mygame.h"
+#include "../../Library/gameutil.h"
 #include "Weapon.h"
 using namespace game_framework;
 Weapon::Weapon()
