+++ Source/Game/mygame_run.cpp
@@ -6,6 +6,7 @@
 #include "../Library/gameutil.h"
 #include "../Library/gamecore.h"
 #include "mygame.h"
+#include "VSclass/VS.h"
 
 using namespace game_framework;
