+++ game.vcxproj
@@ -232,6 +232,7 @@
     <None Include="RES\GAME.RC2" />
   </ItemGroup>
   <ItemGroup>
+    <ClCompile Include="Source\Game\VSclass\Button.cpp" />
     <ClCompile Include="Source\Core\game.cpp" />
     <ClCompile Include="Source\Core\gameDoc.cpp" />
     <ClCompile Include="Source\Core\gameView.cpp" />
@@ -239,20 +240,21 @@
     <ClCompile Include="Source\Core\StdAfx.cpp">
       <PrecompiledHeader Condition="'$(Configuration)|$(Platform)'=='Debug|Win32'">Create</PrecompiledHeader>
     </ClCompile>
-    <ClCompile Include="Source\Game\Enemy.cpp" />
-    <ClCompile Include="Source\Game\Map.cpp" />
+    <ClCompile Include="Source\Game\VSclass\Enemy.cpp" />
+    <ClCompile Include="Source\Game\VSclass\Map.cpp" />
     <ClCompile Include="Source\Game\mygame_initialize.cpp" />
     <ClCompile Include="Source\Game\mygame_over.cpp" />
     <ClCompile Include="Source\Game\mygame_run.cpp" />
-    <ClCompile Include="Source\Game\Passive.cpp" />
-    <ClCompile Include="Source\Game\Pickups.cpp" />
-    <ClCompile Include="Source\Game\Player.cpp" />
-    <ClCompile Include="Source\Game\Weapon.cpp" />
+    <ClCompile Include="Source\Game\VSclass\Passive.cpp" />
+    <ClCompile Include="Source\Game\VSclass\Pickup.cpp" />
+    <ClCompile Include="Source\Game\VSclass\Player.cpp" />
+    <ClCompile Include="Source\Game\VSclass\Weapon.cpp" />
     <ClCompile Include="Source\Library\audio.cpp" />
     <ClCompile Include="Source\Library\gamecore.cpp" />
     <ClCompile Include="Source\Library\gameutil.cpp" />
   </ItemGroup>
   <ItemGroup>
+    <ClInclude Include="Source\Game\VSclass\Button.h" />
     <ClInclude Include="Source\Core\Game.h" />
     <ClInclude Include="Source\Core\gameDoc.h" />
     <ClInclude Include="Source\Core\gameView.h" />
@@ -260,13 +262,14 @@
     <ClInclude Include="Source\Core\Resource.h" />
     <ClInclude Include="Source\Core\StdAfx.h" />
     <ClInclude Include="Source\Game\config.h" />
-    <ClInclude Include="Source\Game\Enemy.h" />
-    <ClInclude Include="Source\Game\Map.h" />
+    <ClInclude Include="Source\Game\VSclass\Enemy.h" />
+    <ClInclude Include="Source\Game\VSclass\Map.h" />
     <ClInclude Include="Source\Game\mygame.h" />
-    <ClInclude Include="Source\Game\Passive.h" />
-    <ClInclude Include="Source\Game\Pickups.h" />
-    <ClInclude Include="Source\Game\Player.h" />
-    <ClInclude Include="Source\Game\Weapon.h" />
+    <ClInclude Include="Source\Game\VSclass\Passive.h" />
+    <ClInclude Include="Source\Game\VSclass\Pickup.h" />
+    <ClInclude Include="Source\Game\VSclass\Player.h" />
+    <ClInclude Include="Source\Game\VSclass\VS.h" />
+    <ClInclude Include="Source\Game\VSclass\Weapon.h" />
     <ClInclude Include="Source\Library\audio.h" />
     <ClInclude Include="Source\Library\gamecore.h" />
     <ClInclude Include="Source\Library\gameutil.h" />
