+++ game.vcxproj.filters
@@ -54,6 +54,13 @@
     <ClCompile Include="Source\Game\mygame_run.cpp">
       <Filter>MyGame</Filter>
     </ClCompile>
+    <ClCompile Include="Source\Game\VSclass\Enemy.cpp" />
+    <ClCompile Include="Source\Game\VSclass\Map.cpp" />
+    <ClCompile Include="Source\Game\VSclass\Passive.cpp" />
+    <ClCompile Include="Source\Game\VSclass\Pickup.cpp" />
+    <ClCompile Include="Source\Game\VSclass\Player.cpp" />
+    <ClCompile Include="Source\Game\VSclass\Weapon.cpp" />
+    <ClCompile Include="Source\Game\VSclass\Button.cpp" />
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="Source\Library\audio.h">
@@ -89,6 +96,14 @@
     <ClInclude Include="Source\Game\config.h">
       <Filter>MyGame</Filter>
     </ClInclude>
+    <ClInclude Include="Source\Game\VSclass\Enemy.h" />
+    <ClInclude Include="Source\Game\VSclass\Map.h" />
+    <ClInclude Include="Source\Game\VSclass\Passive.h" />
+    <ClInclude Include="Source\Game\VSclass\Pickup.h" />
+    <ClInclude Include="Source\Game\VSclass\Player.h" />
+    <ClInclude Include="Source\Game\VSclass\Weapon.h" />
+    <ClInclude Include="Source\Game\VSclass\Button.h" />
+    <ClInclude Include="Source\Game\VSclass\VS.h" />
   </ItemGroup>
   <ItemGroup>
     <ResourceCompile Include="Source\Core\game.rc">
